+++ aiohttp_apigami/__init__.py
@@ -1,6 +1,7 @@
 # mypy: disable-error-code="attr-defined"
 from importlib import metadata
 
+from .compat import flat_error_handler
 from .core import AiohttpApiSpec, OpenApiVersion, setup_aiohttp_apispec
 from .decorators import (
     cookies_schema,
@@ -23,6 +24,7 @@ __all__ = [
     "__version__",
     "cookies_schema",
     "docs",
+    "flat_error_handler",
     "form_schema",
     "headers_schema",
     "json_schema",